refactor: extract duplicated helpers into coms-shared.ts
Move 12 functions from both coms-p2p.ts and coms-net.ts into coms-shared.ts:
ulid, hexFg, isValidHex, fallbackColor, parseFrontmatter, nowIso,
abbreviateModel, findSystemPromptPath, readFrontmatterFromArgv,
readTaskSummary, buildInboundContent, renderTasksPart.

Also extracts FALLBACK_PALETTE, CROCKFORD consts and TasksSummary type.
