+++ extensions/coms/coms-net.ts
@@ -33,6 +33,7 @@ import { Text } from "@earendil-works/pi-tui";
 import { truncateToWidth } from "@earendil-works/pi-tui";
 import { Type } from "typebox";
 import { applyExtensionDefaults } from "./themeMap.js";
+import { ulid, hexFg, isValidHex, fallbackColor, comsParseYamlFrontmatter as parseFrontmatter, nowIso, abbreviateModel, findSystemPromptPath, readFrontmatterFromArgv, readTaskSummary, buildInboundContent, renderTasksPart, FALLBACK_PALETTE, type TasksSummary } from "./coms-shared.js";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import * as os from "node:os";
@@ -53,10 +54,6 @@ const SERVER_URL_ENV = process.env.PI_COMS_NET_SERVER_URL;
 const AUTH_TOKEN_ENV = process.env.PI_COMS_NET_AUTH_TOKEN;
 const PROJECT_ENV = process.env.PI_COMS_NET_PROJECT;
 
-const FALLBACK_PALETTE = [
-	"#72F1B8", "#36F9F6", "#FF7EDB", "#FEDE5D",
-	"#C792EA", "#FF8B39", "#4D9DE0", "#FFAA8B",
-];
 
 // ━━ Shared types (canonical block — mirrored on server) ━━━━━━━━━━━━━━━━━━━
 
@@ -178,116 +175,6 @@ class HttpError extends Error {
 		super(message ?? `HTTP ${status}`);
 		this.status = status;
 		this.body = body;
-	}
-}
-
-// ━━ Helpers — verbatim from coms-p2p.ts (lines 131-210) ━━━━━━━━━━━━━━━━━━━━━━━━
-
-const CROCKFORD = "0123456789ABCDEFGHJKMNPQRSTVWXYZ"; // pragma: allowlist secret
-
-function ulid(): string {
-	const time = Date.now();
-	const rand = crypto.randomBytes(10);
-	let timeStr = "";
-	let t = time;
-	for (let i = 9; i >= 0; i--) {
-		timeStr = CROCKFORD[t % 32] + timeStr;
-		t = Math.floor(t / 32);
-	}
-	let randStr = "";
-	let bits = 0;
-	let value = 0;
-	for (const byte of rand) {
-		value = (value << 8) | byte;
-		bits += 8;
-		while (bits >= 5) {
-			bits -= 5;
-			randStr += CROCKFORD[(value >> bits) & 31];
-		}
-	}
-	return (timeStr + randStr).slice(0, 26);
-}
-
-function hexFg(hex: string, s: string): string {
-	const r = parseInt(hex.slice(1, 3), 16);
-	const g = parseInt(hex.slice(3, 5), 16);
-	const b = parseInt(hex.slice(5, 7), 16);
-	return `\x1b[38;2;${r};${g};${b}m${s}\x1b[39m`;
-}
-
-function isValidHex(hex: string): boolean {
-	return /^#[0-9a-fA-F]{6}$/.test(hex);
-}
-
-function fallbackColor(sessionId: string): string {
-	const h = crypto.createHash("sha256").update(sessionId).digest("hex").slice(0, 8);
-	return FALLBACK_PALETTE[Number(BigInt("0x" + h)) % FALLBACK_PALETTE.length];
-}
-
-function parseFrontmatter(raw: string): { name?: string; description?: string; color?: string; body: string } {
-	const match = raw.match(/^---\n([\s\S]*?)\n---\n([\s\S]*)$/);
-	if (!match) return { body: raw };
-	const frontmatter: Record<string, string> = {};
-	for (const line of match[1].split("\n")) {
-		const idx = line.indexOf(":");
-		if (idx > 0) {
-			const key = line.slice(0, idx).trim();
-			let val = line.slice(idx + 1).trim();
-			if ((val.startsWith('"') && val.endsWith('"')) || (val.startsWith("'") && val.endsWith("'"))) {
-				val = val.slice(1, -1);
-			}
-			frontmatter[key] = val;
-		}
-	}
-	return {
-		name: frontmatter.name,
-		description: frontmatter.description,
-		color: frontmatter.color,
-		body: match[2],
-	};
-}
-
-function nowIso(): string {
-	return new Date().toISOString();
-}
-
-function abbreviateModel(model: string): string {
-	let m = model || "";
-	if (m.startsWith("claude-")) m = m.slice("claude-".length);
-	if (m.length > 14) m = m.slice(0, 14);
-	return m;
-}
-
-function findSystemPromptPath(argv: string[]): string | null {
-	const scan = (flag: string): string | null => {
-		for (let i = 0; i < argv.length; i++) {
-			if (argv[i] === flag && i + 1 < argv.length) {
-				const candidate = argv[i + 1];
-				if (candidate.endsWith(".md")) {
-					try {
-						if (fs.existsSync(candidate) && fs.statSync(candidate).isFile()) {
-							return candidate;
-						}
-					} catch {
-						// fall through
-					}
-				}
-			}
-		}
-		return null;
-	};
-	return scan("--system-prompt") ?? scan("--append-system-prompt");
-}
-
-function readFrontmatterFromArgv(argv: string[]): { name?: string; description?: string; color?: string } {
-	const p = findSystemPromptPath(argv);
-	if (!p) return {};
-	try {
-		const raw = fs.readFileSync(p, "utf-8");
-		const { name, description, color } = parseFrontmatter(raw);
-		return { name, description, color };
-	} catch {
-		return {};
 	}
 }
 
@@ -580,33 +467,6 @@ export default function (pi: ExtensionAPI) {
 		return { ...prev, ...patch };
 	}
 
-	/** Read task summary for this session only (best-effort). */
-	function readTaskSummary(cwd: string, sessionId?: string): { total: number; completed: number; in_progress: number } | null {
-		try {
-			const tasksDir = path.join(cwd, ".pi", "tasks");
-			if (!fs.existsSync(tasksDir)) return null;
-			const candidates = sessionId
-				? [`tasks-${sessionId}.json`, "tasks.json"]
-				: ["tasks.json"];
-			for (const candidate of candidates) {
-				const filePath = path.join(tasksDir, candidate);
-				if (!fs.existsSync(filePath)) continue;
-				try {
-					const data = JSON.parse(fs.readFileSync(filePath, "utf-8"));
-					if (!Array.isArray(data.tasks) || data.tasks.length === 0) continue;
-					let total = 0, completed = 0, in_progress = 0;
-					for (const t of data.tasks) {
-						total++;
-						if (t.status === "completed") completed++;
-						else if (t.status === "in_progress") in_progress++;
-					}
-					return { total, completed, in_progress };
-				} catch { /* skip corrupt file */ }
-			}
-			return null;
-		} catch { return null; }
-	}
-
 	function handleSseEvent(event: string, data: any, _id?: string): void {
 		if (!data || typeof data !== "object") return;
 		switch (event) {
@@ -689,21 +549,6 @@ export default function (pi: ExtensionAPI) {
 		}
 	}
 
-	function buildInboundContent(senderName: string, senderCwd: string, prompt: string, tasks?: Array<{ subject: string; description: string }> | null): string {
-		let content =
-			`[inbound coms-net message from ${senderName} @ ${senderCwd}]\n` +
-			`[reply by writing a normal assistant message \u2014 your turn output is auto-returned to ${senderName}. ` +
-			`DO NOT call coms_net_send/coms_net_await/coms_net_get to reply; that creates a ping-pong loop.]\n\n` +
-			`${prompt}`;
-		if (tasks && tasks.length > 0) {
-			content += `\n\n## Assigned Tasks\n\nThe sender has assigned you these tasks. Use TaskCreate to add them to your task list, then work through them:\n\n`;
-			for (const task of tasks) {
-				content += `- **${task.subject}**: ${task.description}\n`;
-			}
-		}
-		return content;
-	}
-
 	function handleInboundPrompt(data: any): void {
 		const msg_id: string | undefined = data?.msg_id;
 		if (!msg_id || typeof msg_id !== "string") return;
@@ -751,7 +596,11 @@ export default function (pi: ExtensionAPI) {
 			pi.sendMessage(
 				{
 					customType: "coms-net-inbound",
-					content: buildInboundContent(senderName, senderCwd, promptText, tasks),
+					content: buildInboundContent(
+						`[inbound coms-net message from ${senderName} @ ${senderCwd}]\n` +
+						`[reply by writing a normal assistant message — your turn output is auto-returned to ${senderName}. ` +
+						`DO NOT call coms_net_send/coms_net_await/coms_net_get to reply; that creates a ping-pong loop.]`,
+						promptText, tasks),
 					display: true,
 					details: {
 						msg_id,
@@ -1167,15 +1016,7 @@ export default function (pi: ExtensionAPI) {
 			const sep = theme.fg("dim", "  —  ");
 			const purposePart = theme.fg("muted", r.purpose || "");
 
-			let tasksPart = "";
-			if (r.tasks && r.tasks.total > 0) {
-				const pending = r.tasks.total - r.tasks.completed - r.tasks.in_progress;
-				const parts: string[] = [];
-				if (r.tasks.completed > 0) parts.push(theme.fg("success", `${r.tasks.completed}✔`));
-				if (r.tasks.in_progress > 0) parts.push(theme.fg("accent", `${r.tasks.in_progress}◼`));
-				if (pending > 0) parts.push(theme.fg("dim", `${pending}◻`));
-				tasksPart = theme.fg("dim", " ") + parts.join(theme.fg("dim", " "));
-			}
+			const tasksPart = renderTasksPart(r.tasks, theme);
 			const line = " " + swatch + " " + namePart + " " + modelPart + " " + bar + pctPart + tasksPart + sep + purposePart;
 			out.push(truncateToWidth(line, width));
 		}
@@ -1615,7 +1456,11 @@ export default function (pi: ExtensionAPI) {
 				pi.sendMessage(
 					{
 						customType: "coms-net-inbound",
-						content: buildInboundContent(next.sender_name, next.sender_cwd, next.prompt, next.tasks),
+						content: buildInboundContent(
+						`[inbound coms-net message from ${next.sender_name} @ ${next.sender_cwd}]\n` +
+						`[reply by writing a normal assistant message — your turn output is auto-returned to ${next.sender_name}. ` +
+						`DO NOT call coms_net_send/coms_net_await/coms_net_get to reply; that creates a ping-pong loop.]`,
+						next.prompt, next.tasks),
 						display: true,
 						details: {
 							msg_id: next.msg_id,
