+++ extensions/coms/coms-p2p.ts
@@ -19,6 +19,7 @@ import type { ExtensionAPI, ExtensionContext, Theme } from "@earendil-works/pi-c
 import { Text, truncateToWidth } from "@earendil-works/pi-tui";
 import { Type } from "typebox";
 import { applyExtensionDefaults } from "./themeMap.js";
+import { ulid, hexFg, isValidHex, fallbackColor, comsParseYamlFrontmatter as parseFrontmatter, nowIso, abbreviateModel, findSystemPromptPath, readFrontmatterFromArgv, readTaskSummary, buildInboundContent, renderTasksPart, FALLBACK_PALETTE, type TasksSummary } from "./coms-shared.js";
 import * as net from "node:net";
 import * as fs from "node:fs";
 import * as path from "node:path";
@@ -34,10 +35,6 @@ const PING_INTERVAL_MS = Number(process.env.PI_COMS_PING_INTERVAL_MS) || 10_000;
 const KEEPALIVE_INTERVAL_MS = 30_000;
 const LINE_CAP_BYTES = 64 * 1024;
 
-const FALLBACK_PALETTE = [
-	"#72F1B8", "#36F9F6", "#FF7EDB", "#FEDE5D",
-	"#C792EA", "#FF8B39", "#4D9DE0", "#FFAA8B",
-];
 
 // ━━ Types ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
@@ -130,89 +127,11 @@ interface InboundContext {
 	fulfilled: boolean;
 }
 
-// ━━ Helpers ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-const CROCKFORD = "0123456789ABCDEFGHJKMNPQRSTVWXYZ"; // pragma: allowlist secret
-
-function ulid(): string {
-	const time = Date.now();
-	const rand = crypto.randomBytes(10);
-	let timeStr = "";
-	let t = time;
-	for (let i = 9; i >= 0; i--) {
-		timeStr = CROCKFORD[t % 32] + timeStr;
-		t = Math.floor(t / 32);
-	}
-	let randStr = "";
-	let bits = 0;
-	let value = 0;
-	for (const byte of rand) {
-		value = (value << 8) | byte;
-		bits += 8;
-		while (bits >= 5) {
-			bits -= 5;
-			randStr += CROCKFORD[(value >> bits) & 31];
-		}
-	}
-	return (timeStr + randStr).slice(0, 26);
-}
-
-function hexFg(hex: string, s: string): string {
-	const r = parseInt(hex.slice(1, 3), 16);
-	const g = parseInt(hex.slice(3, 5), 16);
-	const b = parseInt(hex.slice(5, 7), 16);
-	return `\x1b[38;2;${r};${g};${b}m${s}\x1b[39m`;
-}
-
-function isValidHex(hex: string): boolean {
-	return /^#[0-9a-fA-F]{6}$/.test(hex);
-}
-
-function fallbackColor(sessionId: string): string {
-	const h = crypto.createHash("sha256").update(sessionId).digest("hex").slice(0, 8);
-	return FALLBACK_PALETTE[Number(BigInt("0x" + h)) % FALLBACK_PALETTE.length];
-}
-
-function parseFrontmatter(raw: string): { name?: string; description?: string; color?: string; body: string } {
-	const match = raw.match(/^---\n([\s\S]*?)\n---\n([\s\S]*)$/);
-	if (!match) return { body: raw };
-	const frontmatter: Record<string, string> = {};
-	for (const line of match[1].split("\n")) {
-		const idx = line.indexOf(":");
-		if (idx > 0) {
-			const key = line.slice(0, idx).trim();
-			let val = line.slice(idx + 1).trim();
-			// strip surrounding quotes for values like color: "#36F9F6"
-			if ((val.startsWith('"') && val.endsWith('"')) || (val.startsWith("'") && val.endsWith("'"))) {
-				val = val.slice(1, -1);
-			}
-			frontmatter[key] = val;
-		}
-	}
-	return {
-		name: frontmatter.name,
-		description: frontmatter.description,
-		color: frontmatter.color,
-		body: match[2],
-	};
-}
-
 function makeEndpoint(sessionId: string): string {
 	if (process.platform === "win32") {
 		return `\\\\.\\pipe\\pi-coms-${sessionId}`;
 	}
 	return path.join(COMS_DIR, "sockets", `${sessionId}.sock`);
-}
-
-function nowIso(): string {
-	return new Date().toISOString();
-}
-
-function abbreviateModel(model: string): string {
-	let m = model || "";
-	if (m.startsWith("claude-")) m = m.slice("claude-".length);
-	if (m.length > 14) m = m.slice(0, 14);
-	return m;
 }
 
 // ━━ CLI flag shape (read via pi.registerFlag/pi.getFlag) ━━━━━━━━━━━━━━━━━━━
@@ -365,34 +284,6 @@ function pruneDeadEntriesAllProjects(): RegistryEntry[] {
 	return out;
 }
 
-/** Read task summary for this session only (best-effort, returns null if unavailable). */
-function readTaskSummary(cwd: string, sessionId?: string): { total: number; completed: number; in_progress: number } | null {
-	try {
-		const tasksDir = path.join(cwd, ".pi", "tasks");
-		if (!fs.existsSync(tasksDir)) return null;
-		// Try session-scoped file first, then project-scoped
-		const candidates = sessionId
-			? [`tasks-${sessionId}.json`, "tasks.json"]
-			: ["tasks.json"];
-		for (const candidate of candidates) {
-			const filePath = path.join(tasksDir, candidate);
-			if (!fs.existsSync(filePath)) continue;
-			try {
-				const data = JSON.parse(fs.readFileSync(filePath, "utf-8"));
-				if (!Array.isArray(data.tasks) || data.tasks.length === 0) continue;
-				let total = 0, completed = 0, in_progress = 0;
-				for (const t of data.tasks) {
-					total++;
-					if (t.status === "completed") completed++;
-					else if (t.status === "in_progress") in_progress++;
-				}
-				return { total, completed, in_progress };
-			} catch { /* skip corrupt file */ }
-		}
-		return null;
-	} catch { return null; }
-}
-
 // ━━ Transport ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
 function probeStaleSocket(endpoint: string): Promise<"in_use" | "stale"> {
@@ -513,44 +404,6 @@ function sendEnvelope(endpoint: string, envelope: Envelope | Pong | { type: stri
 	});
 }
 
-// ━━ System-prompt frontmatter scan ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-function findSystemPromptPath(argv: string[]): string | null {
-	// Prefer --system-prompt (overwrite). Fall back to --append-system-prompt.
-	// These flags are pi-builtin (not extension-registered) so we still scan
-	// argv directly. First match wins per preference order.
-	const scan = (flag: string): string | null => {
-		for (let i = 0; i < argv.length; i++) {
-			if (argv[i] === flag && i + 1 < argv.length) {
-				const candidate = argv[i + 1];
-				if (candidate.endsWith(".md")) {
-					try {
-						if (fs.existsSync(candidate) && fs.statSync(candidate).isFile()) {
-							return candidate;
-						}
-					} catch {
-						// fall through
-					}
-				}
-			}
-		}
-		return null;
-	};
-	return scan("--system-prompt") ?? scan("--append-system-prompt");
-}
-
-function readFrontmatterFromArgv(argv: string[]): { name?: string; description?: string; color?: string } {
-	const p = findSystemPromptPath(argv);
-	if (!p) return {};
-	try {
-		const raw = fs.readFileSync(p, "utf-8");
-		const { name, description, color } = parseFrontmatter(raw);
-		return { name, description, color };
-	} catch {
-		return {};
-	}
-}
-
 // ━━ Default export ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
 export default function (pi: ExtensionAPI) {
@@ -628,17 +481,6 @@ export default function (pi: ExtensionAPI) {
 		try { socket.end(); } catch { /* ignore */ }
 	}
 
-	function buildInboundContent(senderName: string, senderCwd: string, prompt: string, tasks?: Array<{ subject: string; description: string }> | null): string {
-		let content = `[from ${senderName} @ ${senderCwd}]\n\n${prompt}`;
-		if (tasks && tasks.length > 0) {
-			content += `\n\n## Assigned Tasks\n\nThe sender has assigned you these tasks. Use TaskCreate to add them to your task list, then work through them:\n\n`;
-			for (const task of tasks) {
-				content += `- **${task.subject}**: ${task.description}\n`;
-			}
-		}
-		return content;
-	}
-
 	function handlePrompt(socket: net.Socket, env: PromptEnvelope): void {
 		// 1. Hop limit check
 		if (typeof env.hops !== "number" || env.hops >= MAX_HOPS) {
@@ -684,7 +526,7 @@ export default function (pi: ExtensionAPI) {
 			pi.sendMessage(
 				{
 					customType: "coms-inbound",
-					content: buildInboundContent(env.sender_name, env.sender_cwd, env.prompt, env.tasks),
+					content: buildInboundContent(`[from ${env.sender_name} @ ${env.sender_cwd}]`, env.prompt, env.tasks),
 					display: true,
 					details: {
 						msg_id: env.msg_id,
@@ -1129,15 +971,7 @@ export default function (pi: ExtensionAPI) {
 			const sep = theme.fg("dim", "  —  ");
 			const purposePart = theme.fg("muted", r.purpose || "");
 
-			let tasksPart = "";
-			if (r.tasks && r.tasks.total > 0) {
-				const pending = r.tasks.total - r.tasks.completed - r.tasks.in_progress;
-				const parts: string[] = [];
-				if (r.tasks.completed > 0) parts.push(theme.fg("success", `${r.tasks.completed}✔`));
-				if (r.tasks.in_progress > 0) parts.push(theme.fg("accent", `${r.tasks.in_progress}◼`));
-				if (pending > 0) parts.push(theme.fg("dim", `${pending}◻`));
-				tasksPart = theme.fg("dim", " ") + parts.join(theme.fg("dim", " "));
-			}
+			const tasksPart = renderTasksPart(r.tasks, theme);
 			const line = " " + swatch + " " + namePart + " " + modelPart + " " + bar + pctPart + tasksPart + sep + purposePart;
 			out.push(truncateToWidth(line, width));
 		}
@@ -1625,7 +1459,7 @@ export default function (pi: ExtensionAPI) {
 				pi.sendMessage(
 					{
 						customType: "coms-inbound",
-						content: buildInboundContent(next.sender_name, next.sender_cwd, next.prompt, next.tasks),
+						content: buildInboundContent(`[from ${next.sender_name} @ ${next.sender_cwd}]`, next.prompt, next.tasks),
 						display: true,
 						details: {
 							msg_id: next.msg_id,
